requirements.txt
langchain-openai
langchain
langchain-community
python-dotenv
langchain-core
streamlit